Extract the rule explanations builder into utilities.ts so we don't have as much duplicate code.
