+++ src/plugins_rules.ts
@@ -12,37 +12,10 @@ import {
   typescriptRulesExtendEslintRules,
 } from './constants.js';
 import { enableJsPluginRule, isIgnoredPluginRule } from './jsPlugins.js';
-import unsupportedRulesJson from './generated/unsupported-rules.json' with { type: 'json' };
+import { buildUnsupportedRuleExplanations } from './utilities.js';
 
 const allRules = Object.values(rules).flat();
-
-// Convert oxc-style rule keys (e.g. "eslint/no-dupe-args", "react/immutability")
-// to all matching ESLint-style keys, using rulesPrefixesForPlugins for aliases
-// (e.g. react → react-hooks/react-refresh, import → import-x, node → n).
-const unsupportedRuleExplanations: Record<string, string> = {};
-for (const [key, value] of Object.entries(
-  unsupportedRulesJson.unsupportedRules
-)) {
-  const slashIdx = key.indexOf('/');
-  const oxlintPlugin = key.slice(0, slashIdx);
-  const ruleName = key.slice(slashIdx + 1);
-
-  // "eslint/rule-name" → "rule-name" (no prefix in ESLint)
-  if (oxlintPlugin === 'eslint') {
-    unsupportedRuleExplanations[ruleName] = value;
-    continue;
-  }
-
-  // Register under every ESLint prefix that maps to this oxlint plugin.
-  // e.g. for "react/immutability", this registers react/, react-hooks/, react-refresh/.
-  for (const [eslintPrefix, plugin] of Object.entries(
-    rulesPrefixesForPlugins
-  )) {
-    if (plugin === oxlintPlugin) {
-      unsupportedRuleExplanations[`${eslintPrefix}/${ruleName}`] = value;
-    }
-  }
-}
+const unsupportedRuleExplanations = buildUnsupportedRuleExplanations();
 
 /**
  * checks if value is validSet, or if validSet is an array, check if value is first value of it
